refactor: move files inside src folder
