+++ src/components/SelectedLettersText.tsx
