+++ src/components/__tests__/SectionListBasics.test.tsx
@@ -3,10 +3,9 @@ import React from "react";
 import renderer from "react-test-renderer";
 
 // Internal
-import { About } from "./About";
+import { SectionListBasics } from "../SectionListBasics";
 
 it("renders without crashing", () => {
-  const props = {};
-  const rendered = renderer.create(<About {...props} />).toJSON();
+  const rendered = renderer.create(<SectionListBasics />).toJSON();
   expect(rendered).toBeTruthy();
 });
