+++ src/helpers/__tests__/strings.test.ts
