+++ src/helpers/strings.ts
